✅ Properly look for thread name for Android logging tests
